pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cbs"
version = "0.1.0"
description = "Central Bank Swaps — Agentic Scraping Pipeline"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.0",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langchain-google-genai>=2.1",
    "langsmith>=0.2",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "pymupdf>=1.24",
    "pdfplumber>=0.11",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "playwright>=1.58.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
    "types-beautifulsoup4>=4.12",
    "openpyxl>=3.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not benchmark'"
markers = [
    "integration: marks tests that may hit real services (deselect with '-m \"not integration\"')",
    "benchmark: marks accuracy benchmark tests that hit real LLM APIs (run with '-m benchmark')",
]
